+++ src/components/Jobs/Jobs.js
@@ -1,10 +1,6 @@
 import React, { memo } from 'react';
-import { FormattedMessage } from 'react-intl';
 
-import {
-  AccordionSet,
-  Accordion,
-} from '@folio/stripes/components';
+import { AccordionSet } from '@folio/stripes/components';
 
 import {
   PreviewsJobs,
@@ -16,26 +12,8 @@ import css from './Jobs.css';
 export const Jobs = memo(() => (
   <div className={css.jobsPane}>
     <AccordionSet>
-      <Accordion
-        label={(
-          <span data-test-preview-jobs-accordion-title>
-            <FormattedMessage id="ui-data-import.previewJobs" />
-          </span>
-        )}
-        separator={false}
-      >
-        <PreviewsJobs />
-      </Accordion>
-      <Accordion
-        label={(
-          <span data-test-running-jobs-accordion-title>
-            <FormattedMessage id="ui-data-import.runningJobs" />
-          </span>
-        )}
-        separator={false}
-      >
-        <RunningJobs />
-      </Accordion>
+      <PreviewsJobs />
+      <RunningJobs />
     </AccordionSet>
   </div>
 ));