+++ test/bigtest/tests/job-executions/jobs-sort-test.js
@@ -6,7 +6,6 @@ import {
 } from '@bigtest/mocha';
 
 import { sortPreviewJobs } from '../../../../src/components/Jobs/components/PreviewsJobs/sortPreviewJobs';
-import { sortRunningJobs } from '../../../../src/components/Jobs/components/RunningJobs/sortRunningJobs';
 import { JOB_STATUSES } from '../../../../src/utils/constants';
 
 const {
@@ -102,35 +101,10 @@ const previewJobsSorted = [
   },
 ];
 
-const runningJobsUnsorted = [
-  { startedDate: '2018-10-19T18:33:45.000' },
-  { startedDate: '2018-11-22T17:16:13.000' },
-  { startedDate: '2018-10-19T14:32:29.000' },
-  { startedDate: '2018-11-22T15:54:44.000' },
-  { startedDate: '2018-10-19T14:32:29.000' },
-];
-
-const runningJobsSorted = [
-  { startedDate: '2018-11-22T17:16:13.000' },
-  { startedDate: '2018-11-22T15:54:44.000' },
-  { startedDate: '2018-10-19T18:33:45.000' },
-  { startedDate: '2018-10-19T14:32:29.000' },
-  { startedDate: '2018-10-19T14:32:29.000' },
-];
-
 describe('sortPreviewJobs', () => {
   it('sorts correctly', () => {
     const sortFunctionResult = sortPreviewJobs(previewJobsUnsorted);
     const isSortedCorrectly = isEqual(sortFunctionResult, previewJobsSorted);
-
-    expect(isSortedCorrectly).to.be.true;
-  });
-});
-
-describe('sortRunningJobs', () => {
-  it('sorts correctly', () => {
-    const sortFunctionResult = sortRunningJobs(runningJobsUnsorted);
-    const isSortedCorrectly = isEqual(sortFunctionResult, runningJobsSorted);
 
     expect(isSortedCorrectly).to.be.true;
   });